WIP: clean up benchmark script execution: Coremark
Start to clean up the run-benchmarks.sh script. In particular:

* Coremark is updated at the beginning of the script with a `git
submodule`, instead of cloning directly into the repo.
* All benchmarks should be run directly from their benchmark_src/$benchmark
directory, instead of copying the source to the current directory.
* All results should be placed in a new directory, the name of which is
specified when running the script.
* The run_*.sh script functionality will be integrated into
run-benchmarks.sh, since those scripts are just loops that refer to
Makefiles. The process_*.py scripts will probably be kept in the
benchmark_scripts/ directory.

This commit should finish the work for Coremark execution... updates for
Whetstone and Dhrystone to follow.