Finish reorganizing for Whetstone and Dhrystone
Finishes the first round of cleaning up run-benchmarks.sh. The script
now runs entirely from the repo root directory (although it does assume
that the working directory is the repo root).

All the benchmark run logic (compiling, performing multiple runs, etc.)
is now in run-benchmarks.sh, instead of distributed in benchmark_scripts/.

All results are now saved in a new directory created when the script is
run, named "results_$date_$time", where $date and $time are the current
system time, taken using the `date` utility.

A .gitignore has been added to the repository, which is set to ignore this
new results directory.

Hardware monitoring is not working entirely on this commit, but since
monitoring should change in the near future to use iostat instead of ps,
this probably isn't a great issue.

The script now also lists several parameters used to run the benchmarks
before it begins.