Minor bugfixes for run-benchmarks
Fixes a few bugs, and adds another command-line flag.

The setup() function normally tries to pull the Coremark source from Github, but if
git fails for some reason and the source is already downloaded, the script can bypass
this check if passed the "--no-download" flag. This is also documented in the help message.

Also fixed a couple issues with detecting core number (`command` and `nproc` needed to be
wrapped to properly substitute the command) and compilation arguments (-O3, with and alphabetical
"O", not -03 with a zero.)